+++ src/InputGroup/InputGroup.js
@@ -3,16 +3,7 @@ import PropTypes from 'prop-types';
 import { INPUT_GROUP_ADDON_POSITIONS, INPUT_GROUP_TYPES } from '../utils/constants';
 import React, { Component } from 'react';
 
-export const FormGroup = ({ children, ...props }) => {
-    return <div {...props} className='fd-form__group'>{children}</div>;
-};
-FormGroup.displayName = 'FormGroup';
-
-FormGroup.propTypes = {
-    children: PropTypes.node
-};
-
-export class InputGroup extends Component {
+class InputGroup extends Component {
     constructor(props) {
         super(props);
 
@@ -259,6 +250,7 @@ export class InputGroup extends Component {
         }
     }
 }
+
 InputGroup.displayName = 'InputGroup';
 
 InputGroup.propTypes = {
@@ -301,3 +293,5 @@ InputGroup.propDescriptions = {
     numberUpButtonProps: 'Additional props to be spread to the up `<button>` element (for inputType=\'number\').',
     searchButtonProps: 'Additional props to be spread to the `<button>` element.'
 };
+
+export default InputGroup;